lecture 62: add logic to POST user creation information
If the user is loading the page for visualization, the request is of method 'GET'.
If that is the case, display the UserCreationForm, as we used to do.

If the user is sending data, like username and password, in order to sign up for the
site, the request is of method 'POST'. In this case, we need to call the 'create_user'
function from the User.object class, and pass what we want (username and password1 in
in this case).

But this can only happen after we make sure the passwords match each other, so we
first need to check that.

The User.save() function makes sure the data is stored in the database.

Signed-off-by: Daniel Kreling <dkreling@redhat.com>