lecture 64: display information about user when logged in
When user is logged in, we want to see this information (username for now) along
with a link for logout. If it is not logged in, we want to see links for them
to login or sign up for a new account.

We will implement that by using a base.html file with some common code for the
templates. About the user, Django already carries the information along and we only
need to use it by calling "user.is_authenticated", then we add the logic for displaying
the links and the text that will display their username.

On the other templates, we extend the base.html and make sure we indicate where the
content block will inject the contents.

Signed-off-by: Daniel Kreling <dkreling@redhat.com>