refactor for cleanup and django standards
Currently we have Signup class exting View, and this way we implement code that is
already in place inside the CreateView class, so we now will make Signup to extend it.

In order to solve a circular dependency, we use the reverse_lazy function, so it can
be used as a placeholder for the currenttodos url.

CreateView already has a 'get' and a 'post' functions so we do not need to reimplement
them. In our project, we need to override its 'vorm_valid' function to implement the
login, which is not implemented in CreateView.

We also implement the link for Sign Up in the initial page.

reference: https://docs.djangoproject.com/en/3.1/

Signed-off-by: Daniel Kreling <dkreling@redhat.com>