lecture 65: add a logout feature
In the base.html template, we add a button that will log the user out of the
wep application.

To have that in place, we need the request to be POST when we press the logout
link, because if it is a GET, the browsers will log the user out as soon as it
loads the current.html page. It will not do this if it is a POST.

In order to have the request sent as a POST, we need to do it through a form, that
is implemented in the base.html template. Its action will be calling the logoutuser
url in the url.py file, which, in turn, will call the logoutuser function in the
views.py file. This function will then logout the user and load the home.html page.

Signed-off-by: Daniel Kreling <dkreling@redhat.com>