lecture 68: create html structure
Create a url endpoint, an html file and a draft view function (createtodo) to make the
structure in place.

Note that this implementation has only the GET request implemented, so we are able to
see its interface. The POST request (called after we hit the Create button) is still
to be implemented.

Signed-off-by: Daniel Kreling <dkreling@redhat.com>