lecture 70: implement POST request in order to edit a todo
Some features are being introduced, such as the ability to edit a previously created
todo. In order to only have access to the logged user's todos, we pass the request
user to get_object_or_404 call.

The GET method displays a form wich is already filled with the existent todo information
and the POST method will save it.

We also introduce a link to the currenttodos page in order to display the todo's editable
form in a new page with a click.

Signed-off-by: Daniel Kreling <dkreling@redhat.com>